rollback: move rollback-method types and functions to client
RollbackMethod, SubvolumeMode, Ambit enums and the detection functions
are only used by client code. Move them from snapper/AppUtil to
client/snapper/rollback-method to keep the library surface minimal.
