+++ testsuite/Makefile.am
@@ -48,6 +48,10 @@ json_formatter_test_LDADD = -lboost_unit_test_framework ../client/utils/libutils
 
 getopts_test_LDADD = -lboost_unit_test_framework ../client/utils/libutils.la
 
+rollback_method_test_SOURCES = rollback-method-test.cc ../client/snapper/rollback-method.cc
+rollback_method_test_CPPFLAGS = $(AM_CPPFLAGS)
+rollback_method_test_LDADD = -lboost_unit_test_framework ../snapper/libsnapper.la
+
 lvm_utils_test_LDADD = -lboost_unit_test_framework ../snapper/libsnapper.la
 
 range_test_LDADD = -lboost_unit_test_framework ../client/utils/libutils.la
